+++ flask_dance/__init__.py
@@ -3,4 +3,4 @@ from __future__ import unicode_literals
 
 from .consumer import OAuth1ConsumerBlueprint, OAuth2ConsumerBlueprint
 
-__version__ = "2.0.0"
+__version__ = "2.1.0"