+++ CHANGELOG.rst
@@ -3,6 +3,10 @@ Changelog
 
 `unreleased`_
 -------------
+nothing yet
+
+`2.1.0`_ (2019-05-15)
+---------------------
 * Flask-Dance now provides a ``betamax_record_flask_dance`` testing fixture,
   for recording and replaying HTTP requests using Betamax_. See the testing
   documentation for more information.
@@ -308,7 +312,8 @@ Fixed
 .. _#161: https://github.com/singingwolfboy/flask-dance/issues/161
 
 
-.. _unreleased: https://github.com/singingwolfboy/flask-dance/compare/v2.0.0...HEAD
+.. _unreleased: https://github.com/singingwolfboy/flask-dance/compare/v2.1.0...HEAD
+.. _2.1.0: https://github.com/singingwolfboy/flask-dance/compare/v2.0.0...v2.1.0
 .. _2.0.0: https://github.com/singingwolfboy/flask-dance/compare/v1.4.0...v2.0.0
 .. _1.4.0: https://github.com/singingwolfboy/flask-dance/compare/v1.3.0...v1.4.0
 .. _1.3.0: https://github.com/singingwolfboy/flask-dance/compare/v1.2.0...v1.3.0